Maj - algo random
correction odometrie

+++ programme robot/exploration/explorationRandom/algoRandom/algoRandom.ino
@@ -185,7 +185,7 @@ void turnDegreeLeft(int degree_to_turn){
 	}
         
         //Enregistrement du theta
-        theta = ((theta + degree_to_turn) % 360);
+        theta = ((int)theta + degree_to_turn) % 360;
 }
 
 //Fonction tourne a droite de X degres
@@ -199,7 +199,7 @@ void turnDegreeRight(int degree_to_turn){
 	}
         
         //Enregistrement du theta
-        theta = (theta - degree_to_turn) % 360;
+        theta = ((int)theta - degree_to_turn) % 360;
 
 }
